+++ animalbingo_app2/app/assets/stylesheets/application.css
@@ -15,6 +15,6 @@
  * files in this directory. Styles in this file should be added after the last require_* statement.
  * It is generally better to create a new file per style scope.
  *
- *= require_tree .
+ * require_tree .
  *= require_self
  */
