+++ animalbingo_app2/app/controllers/bingo_controller.rb
@@ -2,7 +2,7 @@ class BingoController < ApplicationController
   skip_before_action :require_login
   
   def play
-    @pictures = Dinosaur.order("RAND()").limit(10)
+    @pictures = Fish.order("RAND()").limit(10)
     @image_paths = @pictures.map(&:img) #データの順を同じにするためにpluckではくmapを使用。
     @names = @pictures.map(&:name)
     # binding.pry
