+++ animalbingo_app2/app/controllers/collections_controller.rb
@@ -1,0 +1,12 @@
+class CollectionsController < ApplicationController
+  def index
+    @animals = current_user.animals
+    @animal_imgpaths = @animals.map(&:img) #データの順を同じにするためにpluckではくmapを使用。
+    @animal_imgpaths << 'question101.png'
+    @animal_names = @animals.map(&:name)
+    # binding.pry
+  end
+
+  def show
+  end
+end
