+++ animalbingo_app2/app/javascript/application.js
@@ -2,4 +2,5 @@
 import "@hotwired/turbo-rails"
 import "controllers"
 
-import './custom/bingo_ruret.js'
+
+
