+++ animalbingo_app2/app/views/bingo/play1.html.erb
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>ビンゴ</title>
-    <link href="../../assets/stylesheets/bingo_ruret.css" rel="stylesheet" type="text/css">
+    <%= stylesheet_link_tag 'bingo_ruret', :media => "all" %>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 </head>
 <body>
@@ -29,15 +29,16 @@
             </div>
         </div>
     </div>
-
-    <script src="bingo.js"></script>
-    <script>
-        const images = <%= raw @pictures.to_json %>;
-        const names = <%= raw @names.to_json %>;
+    <% content_for :js do %>
+        <%= javascript_import_module_tag "bingo_ruret" %>
+    <% end %>
+    <script> //javascriptにデータを渡す。
+        images = <%= raw @pictures.to_json %>;
+        names = <%= raw @names.to_json %>;
         //const imagePath = "<%= asset_path('animal/1.png') %>";
-        const imagePaths = <%= raw @image_paths.map { |path| "#{path}" }.to_json %>;
-        const imagePathsArray = <%= raw @image_paths.map { |path| asset_path("#{path}") }.to_json %>;
-  // imagesを使用したJavaScriptのコード
-</script>
+        imagePaths = <%= raw @image_paths.map { |path| "#{path}" }.to_json %>;
+        imagePathsArray = <%= raw @image_paths.map { |path| asset_path("#{path}") }.to_json %>;
+    </script>
+
 </body>
 </html>
