+++ animalbingo_app2/app/views/collections/index.html.erb
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8" /> 
+  <title>collection</title>
+  <%= stylesheet_link_tag 'collection', :media => "all" %>
+</head>
+<body>
+  <div class="container">
+    <div class="bingo">
+         <div class="bingo-heading">
+            <h1>どうぶつ</h1>
+        </div>
+        <div class="card">
+            <div class="outer"  id="outer">
+            </div>
+        </div>
+    </div>
+  </div>
+  <% content_for :js do %>
+    <%= javascript_import_module_tag "collection/index" %>
+  <% end %>
+  <script> //javascriptにデータを渡す。
+        images = <%= raw @animals.to_json %>;
+        names = <%= raw @animal_names.to_json %>;
+       
+        //const imagePath = "<%= asset_path('animal/1.png') %>";
+        imagePaths = <%= raw @animal_imgpaths.map { |path| "#{path}" }.to_json %>;
+        imagePathsArray = <%= raw @animal_imgpaths.map { |path| asset_path("#{path}") }.to_json %>;
+    </script>
+
+</body>
+</html>
