+++ animalbingo_app2/app/views/collections/show.html.erb
@@ -1,0 +1,2 @@
+<h1>Collection#show</h1>
+<p>Find me in app/views/collection/show.html.erb</p>
