+++ animalbingo_app2/app/views/shared/_before_login_header.html.erb
@@ -12,7 +12,7 @@
             <%= link_to "モード選択", "#" %>
         </div>
         <div class='col'>
-            <%= link_to "ログイン", "#" %>
+            <%= link_to "ログイン", login_path %>
         </div>
 
     </div>
