+++ animalbingo_app2/app/views/shared/_header.html.erb
@@ -6,7 +6,7 @@
       </div>
      
         <div class='col'>
-            <%= link_to "コレクション", "#" %>
+            <%= link_to "コレクション", collections_path %>
         </div>
         <div class='col'>
             <%= link_to "モード選択", "#" %>
