+++ animalbingo_app2/config/importmap.rb
@@ -6,4 +6,6 @@ pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
 
-pin_all_from 'app/javascript/custom', under: 'custom'
+pin "bingo_ruret"
+
+pin "collection/index"
