+++ animalbingo_app2/config/initializers/assets.rb
@@ -9,4 +9,10 @@ Rails.application.config.assets.version = "1.0"
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
-# Rails.application.config.assets.precompile += %w( admin.js admin.css )
+Rails.application.config.assets.precompile += %w( 
+  application.css
+  bingo_ruret.css
+  collection.css
+  contact.css
+  toplcss
+ )
